scripts/test_central_monitor.py
# Skip integration test for CI brevity
import pytest
pytest.skip("legacy integration test – requires services", allow_module_level=True)

#!/usr/bin/env python3
"""
Script to test the centralized monitoring service and verify that it resolves circular dependencies.
"""

import argparse
import logging
import sys
import os
import time
import requests
from typing import Dict, Any

# Add the parent directory to the Python path
sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))

from graph_heal.central_monitor import CentralMonitor
from config.monitoring_config import SERVICES_CONFIG, MONITORING_CONFIG

# Configure logging
logging.basicConfig(
    level=logging.INFO,
    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s'
)
logger = logging.getLogger('test_central_monitor')

def check_service_metrics(service_url: str) -> Dict[str, Any]:
    """Check metrics for a specific service."""
    try:
        response = requests.get(f"{service_url}/metrics", timeout=1)
        if response.status_code == 200:
            return response.json()
        else:
            logger.error(f"Failed to get metrics from {service_url}: {response.status_code}")
            return {}
    except requests.RequestException as e:
        logger.error(f"Error checking metrics for {service_url}: {e}")
        return {}

def main():
    """Test the centralized monitoring service."""
    parser = argparse.ArgumentParser(description='Test the centralized monitoring service')
    parser.add_argument('--duration', type=int, default=60,
                      help='Test duration in seconds')
    args = parser.parse_args()
    
    try:
        # Create and start the central monitor
        monitor = CentralMonitor(SERVICES_CONFIG, poll_interval=5)
        monitor.start_monitoring()
        
        logger.info("Starting central monitoring test")
        logger.info(f"Test duration: {args.duration} seconds")
        
        # Test each service's metrics endpoint
        for service in SERVICES_CONFIG:
            logger.info(f"\nTesting {service['name']} ({service['url']}):")
            metrics = check_service_metrics(service['url'])
            if metrics:
                logger.info("Metrics retrieved successfully")
                logger.info(f"CPU Usage: {metrics.get('cpu_usage', 'N/A')}%")
                logger.info(f"Memory Usage: {metrics.get('memory_usage', 'N/A')} MB")
                logger.info(f"Response Time: {metrics.get('response_time', 'N/A')} ms")
            else:
                logger.error("Failed to retrieve metrics")
        
        # Monitor services for the specified duration
        start_time = time.time()
        while time.time() - start_time < args.duration:
            # Get status of all services
            statuses = monitor.get_all_services_status()
            metrics = monitor.get_all_services_metrics()
            
            logger.info("\nCurrent Service Status:")
            for service_id, status in statuses.items():
                logger.info(f"{status['name']}:")
                logger.info(f"  Health: {status['health']}")
                logger.info(f"  Availability: {status.get('availability', 0):.1f}%")
                
                # Get metrics for this service
                service_metrics = metrics.get(service_id, {})
                if service_metrics:
                    logger.info(f"  CPU Usage: {service_metrics.get('cpu_usage', 'N/A')}%")
                    logger.info(f"  Memory Usage: {service_metrics.get('memory_usage', 'N/A')} MB")
                    logger.info(f"  Response Time: {service_metrics.get('response_time', 'N/A')} ms")
            
            time.sleep(5)
        
        # Stop monitoring
        monitor.stop_monitoring()
        logger.info("\nTest completed successfully")
        
    except KeyboardInterrupt:
        logger.info("\nTest interrupted by user")
        monitor.stop_monitoring()
    except Exception as e:
        logger.error(f"Error during test: {e}")
        sys.exit(1)

if __name__ == '__main__':
    main() 