requirements.txt
flask==2.0.1
werkzeug==2.0.1
requests>=2.31.0
matplotlib==3.8.4
numpy==1.26.4
prometheus-client==0.11.0 
psutil 
flask_cors
numexpr<2.12,>=2.8.4
bottleneck<1.6,>=1.3.6
scipy>=1.10
coverage<8,>=7.5.1
pytest-cov<5,>=4.1.0
networkx>=3.2
pyyaml==6.0.1
pandas>=2.0
seaborn==0.12.2
python-louvain==0.16
docker==6.0.1
kubernetes==26.1.0
asyncua==0.9.95
paho-mqtt==2.1.0
scikit-learn==1.5.0
joblib==1.4.2 